packages/memory-mcp-server/pyproject.toml
[project]
name = "bob-memory-mcp-server"
version = "0.1.0"
description = "Memory MCP Server for Bob Agent using Mem0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.1.0",
    "mem0ai>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
bob-memory-server = "memory_server:main"
